Code review: move some tests to test_init
