+++ tests/components/generic/test_config_flow.py
@@ -28,7 +28,7 @@ from homeassistant.components.stream import (
     CONF_RTSP_TRANSPORT,
     CONF_USE_WALLCLOCK_AS_TIMESTAMPS,
 )
-from homeassistant.config_entries import ConfigEntryState, ConfigFlowResult
+from homeassistant.config_entries import ConfigFlowResult
 from homeassistant.const import (
     CONF_AUTHENTICATION,
     CONF_NAME,
@@ -887,35 +887,6 @@ async def test_options_permission_error(
         )
     assert result2["type"] is FlowResultType.FORM
     assert result2["errors"] == {"stream_source": "stream_not_permitted"}
-
-
-@pytest.mark.usefixtures("fakeimg_png")
-async def test_unload_entry(hass: HomeAssistant, setup_entry: MockConfigEntry) -> None:
-    """Test unloading the generic IP Camera entry."""
-    assert setup_entry.state is ConfigEntryState.LOADED
-
-    await hass.config_entries.async_unload(setup_entry.entry_id)
-    await hass.async_block_till_done()
-    assert setup_entry.state is ConfigEntryState.NOT_LOADED
-
-
-async def test_reload_on_title_change(
-    hass: HomeAssistant, setup_entry: MockConfigEntry
-) -> None:
-    """Test the integration gets reloaded when the title is updated."""
-    assert setup_entry.state is ConfigEntryState.LOADED
-    assert (
-        hass.states.get("camera.test_camera").attributes["friendly_name"]
-        == "Test Camera"
-    )
-
-    hass.config_entries.async_update_entry(setup_entry, title="New Title")
-    assert setup_entry.title == "New Title"
-    await hass.async_block_till_done()
-
-    assert (
-        hass.states.get("camera.test_camera").attributes["friendly_name"] == "New Title"
-    )
 
 
 async def test_migrate_existing_ids(
